Created Login Authentication Test
Very Basic Test just checks if there is a username and password that
match the given username and password. It then returns either login
successful or login unsuccessful.

The username I have setup in the database currently is the super
original Test with a password of Test.

There are currently security issues with the entire setup that may need
to be rectified with a web service..